Updated
db - 100% ready
rest_app = 100%
web_app = 80%
testing files = 20%

+++ rest_app.py
@@ -16,36 +16,37 @@ def user(user_id):
         # treating request_data as a dictionary to get a specific value from key
         user_name = request_data.get('user_name')
         users[user_id] = user_name
-        try:
-            db_connector.insert_db(user_id,user_name) #
-            print("User added successfully")
+        does_user_exist = db_connector.select_db(user_id) # check if user exists before adding
+        if does_user_exist is None:
+            db_connector.insert_db(user_id,user_name)
             return {'status': 'ok', 'user_added': user_name}, 200 # status code
-        except pymysql.err.IntegrityError:
-            print("User already exists")
+        else:
             return {'status': 'error', 'reason': 'id already exists'}, 500 # error status code
 
     elif request.method == 'GET':
-        try:
-            user_name = db_connector.select_db(user_id)
+        user_name = db_connector.select_db(user_id)
+        if user_name is not None:
             return {'status': 'ok', 'user_name': user_name}, 200 # status code
-        except:
+        else:
             return {'status': 'error', 'reason': 'id not found'}, 500 # error status code
 
     elif request.method == 'PUT':
         request_data = request.json
         user_name = request_data.get('user_name')
         users[user_id] = user_name
-        try:
+        does_user_exist = db_connector.select_db(user_id)
+        if does_user_exist is not None:
             db_connector.update_db(user_id, user_name)
             return {'user_id': user_id, 'user_updated': users[user_id]}, 200 # status code
-        except:
+        else:
             return {'status': 'error', 'reason': 'no such id'}, 500 # error status code
 
     elif request.method == 'DELETE':
-        try:
+        does_user_exist = db_connector.select_db(user_id)
+        if does_user_exist is not None:
             db_connector.delete_db(user_id)
             return {'user_id': user_id, 'user_deleted': user_id}, 200 # status code
-        except:
+        else:
             return {'status': 'error', 'reason': 'no such id'}, 500 # error status code
